+++ src/Interpreter/Interpreter.cs
@@ -131,6 +131,77 @@ public partial class Interpreter
             ["KEY_F10#"] = 324,
             ["KEY_F11#"] = 389,
             ["KEY_F12#"] = 390,
+            // Numpad keys
+            ["KEY_NUMPAD0#"] = 256,
+            ["KEY_NUMPAD1#"] = 257,
+            ["KEY_NUMPAD2#"] = 258,
+            ["KEY_NUMPAD3#"] = 259,
+            ["KEY_NUMPAD4#"] = 260,
+            ["KEY_NUMPAD5#"] = 261,
+            ["KEY_NUMPAD6#"] = 262,
+            ["KEY_NUMPAD7#"] = 263,
+            ["KEY_NUMPAD8#"] = 264,
+            ["KEY_NUMPAD9#"] = 265,
+            // SHIFT, CTRL, ALT
+            ["KEY_LSHIFT#"] = 340,
+            ["KEY_RSHIFT#"] = 344,
+            ["KEY_LCTRL#"] = 341,
+            ["KEY_RCTRL#"] = 345,
+            ["KEY_LALT#"] = 342,
+            ["KEY_RALT#"] = 346,
+            // Win keys
+            ["KEY_LWIN#"] = 343,
+            ["KEY_RWIN#"] = 347,
+            // Alphabets and numbers
+            ["KEY_0#"] = 48,
+            ["KEY_1#"] = 49,
+            ["KEY_2#"] = 50,
+            ["KEY_3#"] = 51,
+            ["KEY_4#"] = 52,
+            ["KEY_5#"] = 53,
+            ["KEY_6#"] = 54,
+            ["KEY_7#"] = 55,
+            ["KEY_8#"] = 56,
+            ["KEY_9#"] = 57,
+            ["KEY_A#"] = 65,
+            ["KEY_B#"] = 66,
+            ["KEY_C#"] = 67,
+            ["KEY_D#"] = 68,
+            ["KEY_E#"] = 69,
+            ["KEY_F#"] = 70,
+            ["KEY_G#"] = 71,
+            ["KEY_H#"] = 72,
+            ["KEY_I#"] = 73,
+            ["KEY_J#"] = 74,
+            ["KEY_K#"] = 75,
+            ["KEY_L#"] = 76,
+            ["KEY_M#"] = 77,
+            ["KEY_N#"] = 78,
+            ["KEY_O#"] = 79,
+            ["KEY_P#"] = 80,
+            ["KEY_Q#"] = 81,
+            ["KEY_R#"] = 82,
+            ["KEY_S#"] = 83,
+            ["KEY_T#"] = 84,
+            ["KEY_U#"] = 85,
+            ["KEY_V#"] = 86,
+            ["KEY_W#"] = 87,
+            ["KEY_X#"] = 88,
+            ["KEY_Y#"] = 89,
+            ["KEY_Z#"] = 90,
+            // Punctuations
+            ["KEY_SEP#"] = 59, // ;:
+            ["KEY_EQUALS#"] = 61, // =
+            ["KEY_COMMA#"] = 44, // ,
+            ["KEY_MINUS#"] = 45, // -
+            ["KEY_DOT#"] = 46, // .
+            ["KEY_SLASH#"] = 47, // /
+            ["KEY_GRAVE#"] = 96, // `
+            // KEY_LBRACKET#, KEY_RBRACKET#,  KEY_BACKSLASH# 
+            ["KEY_LBRACKET#"] = 91, // [
+            ["KEY_RBRACKET#"] = 93, // ]
+            ["KEY_BACKSLASH#"] = 92, // \
+
             // Mouse
             ["MOUSE_LEFT#"] = 1,
             ["MOUSE_RIGHT#"] = 2,